Remove magicHtml option #379128 (#15)
With #14 , we've also bumped a couple of dependencies. One of it,
`webpack-dev-server` now causes issues while serving a template via `npx
webpack serve --config webpack.config.js --mode development --progress`.
The reason for the failure is the switch to Webpack v5 which introduced
a breaking change* regarding the `magicHtml` option, which has been
removed completly. As we've never used that feature, we can safely
remove the option now.

We should afterwards check in detail whether this solves all issues or
if there are more breaking changes.


*https://github.com/webpack/webpack-dev-server/blob/master/migration-v5.md